fix distributed issues
